// src/payload-types.ts
export interface Payment {
  id: string;
  customerDetails?: {
    title?: string | null;
    firstName?: string | null;
    lastName?: string | null;
    email?: string | null;
    phone?: string | null;
    address?: string | null;
  };
  vehicleDetails?: {
    make?: string | null;
    model?: string | null;
    registration?: string | null;
    stockId?: string | null;
  };
  /**
   * Amount in GBP (pence)
   */
  amount?: number | null;
  currency?: string | null;
  atoaPaymentDetails?: {
    paymentRequestId?: string | null;
    orderId?: string | null;
    customerId?: string | null;
    paymentUrl?: string | null;
    qrCodeUrl?: string | null;
    paymentIdempotencyId?: string | null;
  };
  paymentStatus?: ('PENDING' | 'COMPLETED' | 'FAILED' | 'EXPIRED' | 'CANCELLED') | null;
  statusDetails?: {
    statusUpdateDate?: string | null;
    isoStatus?: {
      code?: string | null;
      name?: string | null;
    };
    errorDescription?: string | null;
  };
  /**
   * Actual amount paid in GBP (pence)
   */
  paidAmount?: number | null;
  /**
   * Tip amount in GBP (pence)
   */
  tipAmount?: number | null;
  customerName?: string | null;
  signatureHash?: string | null;
  reservationMessage?: string | null;
  updatedAt: string;
  createdAt: string;
}